feat: adds freeotp test cases and creates freeotp module

+++ src/util/secret.rs
@@ -1,4 +1,4 @@
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub struct OtpSecret(pub Vec<u8>);
 
 impl serde::Serialize for OtpSecret {
@@ -15,10 +15,10 @@ impl<'de> serde::Deserialize<'de> for OtpSecret {
   where
     D: serde::Deserializer<'de>,
   {
-    Vec::<i8>::deserialize(deserializer).map(|secret| {
+    Vec::<i32>::deserialize(deserializer).map(|secret| {
       let sec = secret
         .iter()
-        .map(|&i| u8::try_from((i as i32 + 256) % 256).expect("!secret"))
+        .map(|&i| u8::try_from((i + 256) % 256).expect("!secret"))
         .collect::<Vec<u8>>();
       OtpSecret(sec)
     })
